StockMarket/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
websockets==12.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
chromadb==0.4.22
openai==1.12.0
python-jose==3.3.0
passlib==1.7.4
pytest==8.0.1
httpx==0.26.0
redis==5.0.1
sentence-transformers==2.3.1
python-multipart==0.0.9
bcrypt==4.1.2 